Accepts console argument to specify action to perform. Two actions available: gdcEtl and makeMdfPcdc
