+++ src/lib/actions.js
@@ -1,0 +1,9 @@
+import gdcEtl from '../gdc/gdcEtl.js';
+import makeMdfPcdc from '../pcdc/makeMdf.js';
+
+const actions = {
+  gdcEtl,
+  makeMdfPcdc,
+};
+
+export default actions;
